+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory+Mandates.swift
@@ -111,6 +111,11 @@ extension PaymentSheetFormFactory {
         return makeMandate(mandateText: mandateText)
     }
 
+    func makeTouchNGoMandate() -> SimpleMandateElement {
+        let mandateText = String(format: String.Localized.touch_n_go_mandate_text, configuration.merchantDisplayName)
+        return makeMandate(mandateText: mandateText)
+    }
+
     func makeNigerianPaymentMethodMandate() -> SimpleMandateElement {
         let mandateText = STPStringUtils.applyLinksToString(
             template: String.Localized.nigerian_payment_method_terms,