+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodParams.swift
@@ -129,6 +129,8 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
     @objc public var scalapay: STPPaymentMethodScalapayParams?
     /// If this is a GoPay PaymentMethod, this contains additional details.
     @objc public var goPay: STPPaymentMethodGoPayParams?
+    /// If this is a Touch 'n Go PaymentMethod, this contains additional details.
+    @objc public var touchNGo: STPPaymentMethodTouchNGoParams?
     /// If this is a GCash PaymentMethod, this contains additional details.
     @objc public var gcash: STPPaymentMethodGCashParams?
     /// If this is a MoMo PaymentMethod, this contains additional details.
@@ -934,6 +936,24 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
         self.metadata = metadata
     }
 
+    /// Creates params for a Touch 'n Go PaymentMethod.
+    /// - Parameters:
+    ///   - touchNGo:       An object containing additional Touch 'n Go details.
+    ///   - billingDetails: Billing information associated with the PaymentMethod.
+    ///   - metadata:       Additional information to attach to the PaymentMethod.
+    @objc
+    public convenience init(
+        touchNGo: STPPaymentMethodTouchNGoParams,
+        billingDetails: STPPaymentMethodBillingDetails?,
+        metadata: [String: String]?
+    ) {
+        self.init()
+        self.type = .touchNGo
+        self.touchNGo = touchNGo
+        self.billingDetails = billingDetails
+        self.metadata = metadata
+    }
+
     /// Creates params for a GCash PaymentMethod.
     /// - Parameters:
     ///   - gcash:       An object containing additional GCash details.
@@ -1059,6 +1079,7 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
             NSStringFromSelector(#selector(getter: sequra)): "sequra",
             NSStringFromSelector(#selector(getter: scalapay)): "scalapay",
             NSStringFromSelector(#selector(getter: goPay)): "gopay",
+            NSStringFromSelector(#selector(getter: touchNGo)): "touch_n_go",
             NSStringFromSelector(#selector(getter: gcash)): "gcash",
             NSStringFromSelector(#selector(getter: momo)): "momo",
             NSStringFromSelector(#selector(getter: ngBankTransfer)): "ng_bank_transfer",
@@ -1526,6 +1547,8 @@ extension STPPaymentMethodParams {
             scalapay = STPPaymentMethodScalapayParams()
         case .goPay:
             goPay = STPPaymentMethodGoPayParams()
+        case .touchNGo:
+            touchNGo = STPPaymentMethodTouchNGoParams()
         case .gcash:
             gcash = STPPaymentMethodGCashParams()
         case .momo: